+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - 'maven-release-plugin'
 - Uploading releases to github
+- Remove logging starter
 
 ### Fixed
 - Do not leak distributionManagement information into the CoffeeNet Applications